requirements.txt
# FastAPI 
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9

# (Rate Limiting)
slowapi>=0.1.9

python-dotenv>=1.0.0

# Google Gemini API
google-generativeai>=0.4.0

# HTTP requests
httpx>=0.26.0
requests>=2.31.0

# Logging و Monitoring
python-json-logger>=2.0.7

# تست
pytest>=8.0.0
httpx>=0.26.0

